Readme fixes, make http or https listener optional
Fix some config file options in readme.
Allow to deactivate either http or https port and only use one.
Hostnames in URL can be prefixed with https: or http: to specify
the protocol to use for the destination. This to allow access to the
proxy by http but still have access to https servers (or the other way),
e.g.:
  http://localhost:8080//https:download.rockylinux.org/pub/rocky/...rpm